requirements.txt
opencv-python
numpy
pandas
matplotlib
pillow
easydict